Try to implement exponentiation
Fix a bug in vnormalize() which could even remove the very last limb of
a multiprecision integer.

Implement a bit-shift to the right.

Add a note that the multiplication routine does not allow the same
reference to be passed as both arguments.

Start implementing the pow() function by squaring - does not work yet.