Exponentiation works
At least the non-modular one.

Improve vmult to support squaring as well and remove its restriction
about not allowing references to the same variable for both arguments.

Also implement an optimization about swapping the arguments internally,
favoring the longer argument for addition in the inner loop and the
shorter argument for the outer loop (which is potentially more expensive
to calculate).