Trust the calculated mantissa width
I previously distrusted the result 53 when the script initially
calculated the mantissa bit width for IEEE-754 binary 64-bit floating
point numbers, knowing that only 52 bits are actually being stored for
the mantissa.

I made the mistake of assuming that the implied 53th bit (always "1")
would be useless in integer calculations.

However, this is not actually a problem: It the 53th bit needs to be
"1", it will be due to its implicit character.

And if it needs to be "0", the exponent implicitly shifts the mantissa
to the right in such a way that the implied "1"-bit matches the first
"1"-bit that is actually present somewhere within the number (the value
0 is handled as a special case anyway).

Therefore, there are actually 53 bits available, even if the mantissa
can store only 52 bits. But the exponent helps "generating" the missing
bit when it is needed.

Therefore use the calculated width, and do no longer shorten it by one
bit because of the implicit MSB.