Check for valid numeric ranges for arguments
Because division is not yet implemented, we cannot just reduce some
large value modulo the specified modulus.

Instead, the modulus reduction is done by a single subtraction at a
time, which only works if all involved arguments other than the modulus
are numerically smaller than the modulus.

This has not been checked until now.

Implement a comparison routine in order to allow checking for the
allowed numeric range, and display error messages if a range violation
has been found.