New branch with UUID 8tmif357lnmk9620er12jnujp
Initial branch nickname: modexp-work

Initial check-out location $PROJECTS/modexp/work

This is the development branch for master branch "modexp" with UUID
2fg2id73y3ayyhu66ungjwu10.

The scripts in the master branch share many identical snippets. This
branch stores those snippets separately, avoiding duplicated code.

It also contains *.order files which describe how to reassemble the
scripts from the snippets, and a copy of the reassembled scripts as
well.

The scripts necessary for splitting and reassembling are also available
here.

The master branch "modexp" is synchronzied with this branch, but will
only contain the copies of the reassembled scripts. All other files of
this branch which are only of interest for code refactoring will be
deleted there.

End users only need the master branch. This modexp-work branch is only
for developers.